etl/requirements.txt
pandas
pyarrow
adlfs
azure-storage-file-datalake
azure-identity
python-dotenv
requests
